refactor: Moved container data from pages dir to container dir
- Moved SearchTrack to container dir
- Moved TrackDetails to container dir
